Improve display of websites not specifying background color
The `index.html` of the front-end specifies `#000` as background color,
however, it seems like various websites found 'in the wild' don't
specify their own background color and thus default to `transparent`. As
a result of this, they tend to become mostly unreadable when displayed
using `openWebsite` since they often assume that they get displayed on a
white background. Setting white as background color on the container
element of the `<iframe>` works around this issue.