Fixing camera led always on
This commit fixes a race condition in the "getUserMedia" call that could lead to the webcam being in an always on state.

If we open and close the webcam really quickly, the camera close was sometimes passing BEFORE the open was fully resolved (because getUserMedia can take a few 100ms to answer properly).
As a result, the webcam would stay open.
To solve this, we are putting all calls to getUserMedia in a resolve chain (`currentGetUserMediaPromise`)

Closes #2149