Adding a check of the front/pusher version
When connecting, we will now verify that the pusher and the front
have a compatible API version.
If the front is not compatible with the pusher (because the pusher was updated and the front is still in the old version),
a message is displayed to propose updating to the new version.

The API version is actually a hash of the proto messages files + the Zod API files.
So if anything changes there, the users will have to reload their WorkAdventure.