Moving the START_ROOM_URL to the pusher
The START_ROOM_URL environment variable is now applied to the Pusher and not to the Front.
Unlike before, it is only applied when no AdminAPI is set.

This means users of AdminAPI can now decide where the "/" route will lead
without having to reconfigure the front container (and therefore
that depending on your user, you can redirect to one place or another)

WARNING! This is a breaking change in the containers configuration as
the environment variable must be moved in all production setups from the front to the pusher.