Trying to fix "reconnecting" screen issue
In rare circumstances, the "reconnect" screen stays displayed (even if behind the scene, the reconnection has successfully happened).
We try here to fix this by forcing the removal of the errorScreen (even if the scene is not sleeping).

Also, we make a difference in the user message between a "reconnection failed" and a "connection did not happen yet, even after 1 second"