Fixing issue with local/unreachable maps and BBB/Jiti meetings
In v1.11, we are now correctly handling the reading of the jitsiRoomAdminTag property (or bbbMeetingAdminTag respectively).
We read this from the back server (because reading it from the front could have allowed hackers to lie in the tag).

However, there is an issue with maps that are not readable from the back (like "localhost" or any map behind a firewall).
This issue fixes the problem by bypassing reading the jitsiRoomAdminTag or bbbMeetingAdminTag instead of crashing.