Setting up a query/answer system between front and back
We are adding a new QueryMessage and the mapping AnswerMessage in the protobuf file.
Those query and answer messages can contain a number of possible underlying messages.

We set up some login in RoomConnection, in pusher and in the back to:

- send a query (front)
- forward the query (in the pusher)
- answer the query (and possibly catch any exception and turn them into error answers)
- get the answer in the front, match it to the query id and answer the promise.

With this, we can make queries in RoomConnection that return a promise that will either resolve or fail.

We use this case in Jitsi rooms, to fetch the JWT token.