Trying to implement more robust ping/pong mechanism
At times, it seems that the pusher can kill connexion with the front saying there is no pong received.
Previous implementation was waiting for a pong message every 12 seconds (not related to ping messages sent)
Here, we try to solve this by only creating the "pong" timeout when the ping message was sent.