Fixing jitsiRoomAdminTag not handled when uppercase jitsi room name
The previous fix trying to fix broken "jitsiRoomAdminTag" did only part of the job.

We tried to remove the prefix added by the front to make the Jitsi room unique

But the "jitsiRoom" is actually slugified on the front (in addition to a prefix being added)

As a result, if the jitsiRoom property contained an uppercase or a special character, the back could not find this jitsiRoom property.

This commit fixes everything by computing the exact same slug on the front and the back.
To do so, a new "shared-utils" library is added in "/lib" to contain small utility functions that are common to several containers.