Migrating from retry-axios to axios-retry
Axios-retry is better maintained.
Also, this forces us indirectly to update axios on all containers (not only play) because of a Typescript typing issue (Typescript in a monorepo is using the types of the wrong Axios package)
So this commit ALSO updates Axios for ALL containers to the latest version.