Migrating back to ts-proto
Just like we did with the pusher a few days ago, we are migrating the back to ts-proto.
We are now almost grpc-js free, which is awesome because code generated with ts-proto is way more easy to work with.

The last thing remaining will be switching the service signature from grpc-js to "generic".