Adding a temporary test mode to debug Chrome heavy throttling
Chrome can enable heavy throttling on some tabs, which increases the setTimeout/setInterval
calls to a minimum of 1 second (lite) or 1 minute (heavy).

Here, we are monkey-patching the setTimeout and setInterval functions
to detect if they were throttled. If yes, we are also adding a console.trace() call
to know what function was really throttled.