Do not prevent people from embedding links
Even if the server-side check says a link cannot be used in an iframe, there are very special cases where it is possible.

For instance, in the case of an "internal link" (accessible only from within a company network), the query make from the WorkAdventure server can return a 200 HTTP OK not embeddable link. For instance, if the HTTP 200 contains a page explaining you don't have access to the resource (this has been actually seen today).
This commit makes it possible to embed a link (and not open a new tab) even if the server-side check says it is not possible (to account for those special cases).