Fixing exits in map editor
The implementation of exits in the map editor was flawed because it was based on an exit generated from the WAM url.
It was trying to guess the room URL by prepending it with /~/. This is only true in self-hosted mode. With an admin the mapping between room URL and WAM URL becomes inpredictable.
We are now asking the admin for the list of all rooms.

WARNING! BREAKING CHANGE!

The expected answer from the admin for /api/room/sameWorld has changed.