Fallback in case of Overconstrained deviceId
If a specific camera or microphone cannot be fetched, we now automatically fallback trying to get the default camera or microphone.

Note: this was already the case somehow because the device list is checked and we never request a device not part of the device list.
Yet, in some specific cases (I have a hard time understanding), the device list would always be empty.

This should fix the issue.