Improving fluidity on loaded maps
If there are 100 people packed at the same position in the map, when you move
next to the zone that contains those people, you will suddenly have to
create 100 character instances.

It turns out the creation of a "Text" Phaser object can take up to 16ms.
We are now building this Text object in a setTimeout. Because requestAnimationFrame has the precedence
over setTimeout, this will avoid having freezes. Instead, Wokas will be displayed
without a text for a few milliseconds and the text will be displayed just after, when it is ready.