Trying to detect if no video stream received
We are using HTMLVideoElement.requestVideoFrameCallback to detect the first frame displayed.
If the first frame does not arrive 3 seconds after the constraint store said we should hae video, then we have an issue with the video stream.

For now, we simply display a warning and send an event to the analytics.