Adding support for sending/listening to bubble chat messages in the scripting API
The chat API has been extended to be able to send messages to remote users, but also to listen to messages from remote users.

In the process: a number of types have been added (there is still work to do to put types of events from the chat and play containers in common).
The proximity meeting API now waits for a connection to be established before being triggered. This way, when the proximity API is triggered, we know we can send a message right away via the chat and it will be received.