Fixing detection of multiple users in a bubble in the scripting API (#3602)
The proximityMeeting feature of the scripting API was unable to trigger a call to "onJoin" with several users.
Instead, it would always tell there is 1 user in the bubble, and the participantJoined event could be triggered after or before, a bit randomly.

Now, this is fixed. When joining a bubble with several people, only "onJoin" is triggered, passing in parameter all the bubble participants properly.