Merging master in dev (#3906)
* Fixing close button not displayed in UI

When switching a map, we call the coWebsiteManager cleanup method. This method was unsubscribing stores that were only set once at startup.
This was breaking some features, in particular the fact that the closable button would be automatically hidden.

We are no longer unsubscribing to this store since the subscription should have the same lifetime has the coWebsiteManager singleton (i.e. it should always remain active).

* Switching the build of the desktop app to Node 16

Node 14 is no longer available.

* Fixing bypassing collidion by moving with mouse close to collision

When moving close to a collide tile with the mouse, the Woka would step parially on the collided block.
This would further allow to pass through the wall completely.
We fix this by making sure a Woka hitbox is 100% within the tile targetted (which is non collidable).

* Fixing variable optimization

Previously, if a variable was not changed, on calling `saveVariable`, we would not do anything.
This issue is that if the variable was an object or an array, pushing a new value to the array or modifying a key of the object would not change the reference of the object.
As a result, saving a modified object would not work.

We fix this by disabling the check if we are only modifying an object or array.

* Fixing the drag the map in explorer map to be pixel perfect

Now, the clicked portion of the map remains exactly under the cursor when the map is moved in explorer mode.

* Smoothly slowing the camera velocity when user release click in explorer mode

When the user moves the screen in explorer mode, the velocity is now
kept and slowly decreased to the target location.
This adds a nice extra animation and is fun to play with.

* Automatically marking the scene dirty when the camera moves.

This makes the pan animation really smooth.

* Fixing pan issue when several pans were chained

* Adding some resistance in the zoom out effect

When switching to exploration mode, the zoom will now "resist" and you will have to force to go to exploration mode (renamed "map overview")
When zooming back, you will go automatically to normal mode. You will also face some resistance.

I removed the exploration mode warning popup because the resistance should be enough of a warning and is more natural.
Also, I removed some text in the right side panel, and the panel is closed by default in exploration mode. Only the search feature remains
and you can access it by clicking the magnifying glass.

* Implementing smooth zoom in and out

When zooming in and out on a mouse with wheel, the zoom in and out was done by steps, violently.
We implemented a smooth transition to have something less aggressive to the eye.

* Fixing exploration E2E test

* Setting up a resistance "wall" for zoom in-out.

The wall can be "in-place" or "broken". When the wall is in place, it is NOT possible to pass the resistance zone.
We do this by altering the zoom factor to make it slower as we are getting closer to the resistance zone end.

When we get out of the resistance zone OR if we are in the resistance zone but zoom towards the start of the zone,
we break the wall for 10 seconds.

* Making zoom faster on Mac because MacOS trackpad is slow.

* Changing the way we compute zoomFactor from deltaY.
The new formula can never result in a negative zoomFactor.

* When zooming out to the point where one game pixel is smaller than one screen pixel,
we ensure that the canvas size and height is exactly the canvas width and height (so that 1 canvas pixel === 1 screen pixel)

* Destroying wall automatically after 2 seconds in resistance zone.
Adding a clamp to max wheel values.

* camera.pan has the nasty habit to play poorly when the zoom is modified at the same time.
Removing camera.pan() and implementing our own for camera speed.

Also, fixing an issue where the camera center would be incorrectly set after
a zoom / pan.

* Fixing a bug where in max zoom in mode, we could not move with mouse.
Allowing to zoom in in exploration mode if not centered on Woka.
Removing resistance wall on zoom in.
Lowering resistance on zoom in.

* Fixing zoom factor attenuation.

* Stopping using scroll offset. Instead, focusing on a virtual point with "startFollow"

* Centering the webcam on the world when clicking on the "editor" button to enable zoom out.

* [WIP] Adding a way to typecheck Scripting API variables

The new WA.state.buildValidatedState(zodObject) function
returns a fully typed state.

* Using generic types for the player state related objects

The ultimate goal is to allow proper typing of variables.

* Switching to creating an empty PublicPlayerState interface than can be extended from outside.
Removing the generics.

* Adding the PrivatePlayerState and RoomState "extendable" interfaces.

* Adding documentation for state interfaces and exporting the interfaces in the package.

* Revert "Fix to send event api (#3862)"

This reverts commit acf8fff35bcd59a94aa49d8e46ec975e7161bcf8.

* Removing the black outline around entities in explorer mode

The black outline around entities as a dreadful effect on performances when zooming out.

* Improving world list

Fixing issues in self-hosted world where:
- the active world would not be detected
- the metadata (thumbnail in particular) would not be read

Displaying a default thumbnail if the image thumbnail of the map fails loading properly.

Adding a small animation on mouse hover.

* Fixing Helm release chart deployment

An error in the way we replace the version number in the Chart file
prevents the chart file from being properly generated in the CI.

* Fix focus on object/area in map explorer

When hovering on an object or area, the viewport would no more center on it.
This is now fixed.

* Adding a smooth animation when centering on entity / area